refactor : useHabitContext 훅 분리
- 커스텀 훅은 별도 폴더에 모아두는 형태를 선호함
